requirements.txt
fastapi
pydantic
uvicorn
ollama
sqlite3
faiss-cpu  # Add for FAISS support (use faiss-gpu if you have a GPU)
numpy